Fixed bug: Rendering and reverting the same selection would result in loss of Markdown
Pretty major change in the render-revert logic. It now looks for
wrapper inside the current selection, as well as wrappers above it. So
it's now more aggressive about reverting, but that's good: a) failing
to revert when we should results in loss of Markdown, which is
equivalent to "loss of user data"; b) it probably better represents the
user's intention. Has the cool side-effect of being able to revert
multiple chunks with the same big selection.