use codecogs for LaTeX rendering by default
Note that the changes in the translation files is mostly just churn. The non-English locales' JSON files weren't properly getting English fallbacks added to them; I did that and sorted the keys on output, to prevent future churn. But the difference in order and whitespace bloated the diff. In the future this shouldn't happen.